+++ eva/src/components/modalSettings.vue
@@ -1609,6 +1609,7 @@ export default {
                 'underline',
                 'onButton',
                 'pinned',
+                'numberPerDigit',
               ];
               if (propsToFalse.includes(item)) {
                 localOptions[item] = false;